api/requirements.txt
reportlab>=4.1.0
pypdf>=4.0.0
Pillow>=10.0.0
